requirements.txt
httpx>=0.27
PyJWT>=2.8
pyyaml>=6.0
playwright>=1.44
